d/not-installed: don't package unneeded nftables.service.8 and main.nft
Since 1.1.5, upstream has included a systemd service file, related man-page and
example rule-set, By default the service file is not installed, but the man-page
and rule-set are.  However, since the Debian package provides its own service
file, the upstream one is not required, so we don't package the man-page or
rule-set, either.

Signed-off-by: Jeremy Sowden <azazel@debian.org>